create src/data, notebooks and src/visualization
