+++ requirements-macos.txt
@@ -1,0 +1,76 @@
+appnope==0.1.4
+asttokens==3.0.0
+certifi==2025.4.26
+charset-normalizer==3.4.2
+cloudpickle==3.1.1
+comm==0.2.2
+contourpy==1.3.2
+coverage==7.8.2
+cycler==0.12.1
+debugpy==1.8.14
+decorator==5.2.1
+exceptiongroup==1.3.0
+executing==2.2.0
+farama-notifications==0.0.4
+filelock==3.18.0
+fonttools==4.58.0
+fsspec==2025.5.1
+gymnasium==1.1.1
+idna==3.10
+imageio==2.37.0
+imageio-ffmpeg==0.6.0
+importlib-metadata==8.7.0
+iniconfig==2.1.0
+ipykernel==6.29.5
+ipython==9.2.0
+ipython-pygments-lexers==1.1.1
+jedi==0.19.2
+jinja2==3.1.6
+jupyter-client==8.6.3
+jupyter-core==5.7.2
+kiwisolver==1.4.8
+markupsafe==3.0.2
+matplotlib==3.10.3
+matplotlib-inline==0.1.7
+moviepy==2.2.1
+mpmath==1.3.0
+nest-asyncio==1.6.0
+networkx==3.4.2
+numpy==2.2.6
+packaging==25.0
+pandas==2.2.3
+parso==0.8.4
+pexpect==4.9.0
+pickleshare==0.7.5
+pillow==11.2.1
+platformdirs==4.3.8
+pluggy==1.6.0
+proglog==0.1.12
+prompt-toolkit==3.0.51
+psutil==7.0.0
+ptyprocess==0.7.0
+pure-eval==0.2.3
+pygame==2.6.1
+pygments==2.19.1
+pyparsing==3.2.3
+pytest==8.3.5
+python-dateutil==2.9.0.post0
+python-dotenv==1.1.0
+pytz==2025.2
+pyzmq==26.4.0
+requests==2.32.3
+setuptools==80.8.0
+six==1.17.0
+stack-data==0.6.3
+sympy==1.14.0
+torch==2.7.0
+torchaudio==2.7.0
+torchvision==0.22.0
+tornado==6.5.1
+tqdm==4.67.1
+traitlets==5.14.3
+typing-extensions==4.13.2
+tzdata==2025.2
+urllib3==2.4.0
+wcwidth==0.2.13
+zipp==3.21.0
