+++ requirements.txt
@@ -1,84 +1,74 @@
-asttokens==2.4.1
-certifi==2024.8.30
-charset-normalizer==3.4.0
-cloudpickle==3.1.0
-comm==0.2.2
-contourpy==1.3.0
-coverage==7.6.5
-cycler==0.12.1
-debugpy==1.8.7
-decorator==4.4.2
-exceptiongroup==1.2.2
-executing==2.1.0
-Farama-Notifications==0.0.4
-filelock==3.16.1
-fonttools==4.54.1
-fsspec==2024.10.0
-gymnasium==1.0.0
-idna==3.10
-imageio==2.36.0
-imageio-ffmpeg==0.5.1
-importlib_metadata==8.5.0
-iniconfig==2.0.0
-ipykernel==6.29.5
-ipython==8.29.0
-jedi==0.19.1
-Jinja2==3.1.4
-jupyter_client==8.6.3
-jupyter_core==5.7.2
-kiwisolver==1.4.7
-MarkupSafe==3.0.2
-matplotlib==3.9.2
-matplotlib-inline==0.1.7
-moviepy==1.0.3
-mpmath==1.3.0
-nest-asyncio==1.6.0
-networkx==3.4.2
-numpy==2.1.2
-nvidia-cublas-cu12==12.4.5.8
-nvidia-cuda-cupti-cu12==12.4.127
-nvidia-cuda-nvrtc-cu12==12.4.127
-nvidia-cuda-runtime-cu12==12.4.127
-nvidia-cudnn-cu12==9.1.0.70
-nvidia-cufft-cu12==11.2.1.3
-nvidia-curand-cu12==10.3.5.147
-nvidia-cusolver-cu12==11.6.1.9
-nvidia-cusparse-cu12==12.3.1.170
-nvidia-nccl-cu12==2.21.5
-nvidia-nvjitlink-cu12==12.4.127
-nvidia-nvtx-cu12==12.4.127
-packaging==24.1
-pandas==2.2.3
-parso==0.8.4
-pexpect==4.9.0
-pickleshare==0.7.5
-pillow==11.0.0
-platformdirs==4.3.6
-pluggy==1.5.0
-proglog==0.1.10
-prompt_toolkit==3.0.48
-psutil==6.1.0
-ptyprocess==0.7.0
-pure_eval==0.2.3
-pygame==2.6.1
-Pygments==2.18.0
-pyparsing==3.2.0
-pytest==8.3.3
-python-dateutil==2.9.0.post0
-pytz==2024.2
-pyzmq==26.2.0
-requests==2.32.3
-setuptools==75.3.0
-six==1.16.0
-stack-data==0.6.3
-sympy==1.13.1
-torch==2.5.1
-tornado==6.4.1
-tqdm==4.66.6
-traitlets==5.14.3
-triton==3.1.0
-typing_extensions==4.12.2
-tzdata==2024.2
-urllib3==2.2.3
-wcwidth==0.2.13
-zipp==3.20.2
+asttokens
+certifi
+charset-normalizer
+cloudpickle
+comm
+contourpy
+coverage
+cycler
+debugpy
+decorator
+exceptiongroup
+executing
+Farama-Notifications
+filelock
+fonttools
+fsspec
+gymnasium
+idna
+imageio
+imageio-ffmpeg
+importlib_metadata
+iniconfig
+ipykernel
+ipython
+jedi
+Jinja2
+jupyter_client
+jupyter_core
+kiwisolver
+MarkupSafe
+matplotlib
+matplotlib-inline
+moviepy
+mpmath
+nest-asyncio
+networkx
+numpy
+packaging
+pandas
+parso
+pexpect
+pickleshare
+pillow
+platformdirs
+pluggy
+proglog
+prompt_toolkit
+psutil
+ptyprocess
+pure_eval
+pygame
+Pygments
+pyparsing
+pytest
+python-dateutil
+pytz
+pyzmq
+requests
+setuptools
+six
+stack-data
+sympy
+tornado
+--find-links https://download.pytorch.org/whl/cpu
+torch
+torchvision
+torchaudio
+tqdm
+traitlets
+typing_extensions
+tzdata
+urllib3
+wcwidth
+zipp
