+++ src/data/generate_random.py
@@ -8,7 +8,7 @@ import numpy as np
 import numpy.typing as npt
 import pandas as pd
 
-from data_script.prfndim_utils import expr_eval_safe
+from src.data.prfndim_utils import expr_eval_safe
 from prfndim.prfndim import C, Expr, OverflowError, P, R, S, Z
 
 BATCH_SIZE = 200
